==> nbs/05_train_longvicuna__collect_data.ipynb <==
from datasets import load_dataset
import pandas as pd
from sklearn.model_selection import train_test_split
from markdownify import markdownify
import json
import os

RANDOM_STATE = 42

## Inidivudual datasets preparation

We will use the following datasets:
- OpenAssistant + Alpaca + Gpt4All to keep the original instruction chat-like behaviour
- a few other datasets to make model be able to operate the long-term memory
    - booksum - to generate literature summaries
    - govreport - same, but for formal texts
    - qasper - to operate a memory for QA tasks

### Open Assistant

dataset_oasst1 = load_dataset("OpenAssistant/oasst1")
dataset_oasst1

def _oasst1_parent2children(df_oasst1):
    parent_id_2_children = {}
    for _, row in df_oasst1.iterrows():
        if row["parent_id"] not in parent_id_2_children:
            parent_id_2_children[row["parent_id"]] = []
        parent_id_2_children[row["parent_id"]].append(row["message_id"])
    return parent_id_2_children

def _oasst1_tree(parent_id_2_children, root):
    children = parent_id_2_children.get(root, [])
    result = {}
    for child_id in children:
        result[child_id] = _oasst1_tree(parent_id_2_children, child_id)
    return result

def _oasst1_tree2list(tree):
    if len(tree) == 0:
        return [[]]
    result = []
    for child_id, subtree in tree.items():
        for row in _oasst1_tree2list(subtree):
            result.append([child_id] + row)
    return result

def oasst1_2_messages(df_oasst1):
    dialogue_message_ids = _oasst1_tree2list(
        _oasst1_tree(
            _oasst1_parent2children(df_oasst1),
            None
        )
    )
    message_id2index = {mid: i for i, mid in df_oasst1["message_id"].to_dict().items()}
    df = df_oasst1[["text", "role"]]
    indices = [
        [
            message_id2index[mid]
            for mid in message_ids
        ]
        for message_ids in dialogue_message_ids
    ]
    return indices, df

df_oasst1_train = pd.DataFrame(dataset_oasst1["train"])
df_oasst1_train = df_oasst1_train.loc[df_oasst1_train["review_result"].fillna(False)]
df_oasst1_train.head()

df_oasst1_validation = pd.DataFrame(dataset_oasst1["validation"])
df_oasst1_validation = df_oasst1_validation.loc[df_oasst1_validation["review_result"].fillna(False)]
df_oasst1_validation.head()

oasst1_indices_train, df_oasst1_texts_train = oasst1_2_messages(df_oasst1_train)
df_oasst1_texts_train.loc[oasst1_indices_train[0]]

oasst1_indices_validation, df_oasst1_texts_validation = oasst1_2_messages(df_oasst1_validation)
df_oasst1_texts_validation.loc[oasst1_indices_validation[0]]

print(df_oasst1_texts_validation.loc[oasst1_indices_validation[0]]["text"].values[-1])

### Alpaca

df_alpaca = pd.DataFrame(load_dataset("tatsu-lab/alpaca")["train"])
df_alpaca.head()

df_alpaca["input_message"] = df_alpaca.apply(
    lambda row: f"Below is an instruction that describes a task, paired with an input that provides further context. " + \
        "Write a response that appropriately completes the request.\n\n" + \
        "### Instruction:\n" + \
        f"{row['instruction']}\n" + \
        "\n" + \
        "### Input:\n" +
        f"{row['input']}",
    axis=1
)
df_alpaca["output_message"] = df_alpaca["output"]
df_alpaca = df_alpaca[["input_message", "output_message"]]

df_alpaca_texts = []
alpaca_indices = []
for _, row in df_alpaca.iterrows():
    df_alpaca_texts.append({"role": "prompter", "text": row["input_message"]})
    df_alpaca_texts.append({"role": "assistant", "text": row["output_message"]})
    alpaca_indices.append([len(df_alpaca_texts) - 2, len(df_alpaca_texts) - 1])
df_alpaca_texts = pd.DataFrame.from_records(df_alpaca_texts)

alpaca_indices_train, alpaca_indices_validation = train_test_split(
    alpaca_indices,
    random_state=RANDOM_STATE,
    test_size=0.05
)

df_alpaca_texts.loc[alpaca_indices_train[0]]

df_alpaca_texts.loc[alpaca_indices_train[1]]

### GPT4All

df_gpt4all = pd.DataFrame(load_dataset("nomic-ai/gpt4all_prompt_generations")["train"])
df_gpt4all.head()

df_gpt4all = df_gpt4all.loc[~df_gpt4all["response"].str.contains("As an AI language")]
df_gpt4all.head()

def _markdownify_if_possible(text):
    try:
        return markdownify(text)
    except:
        return None

df_gpt4all["prompt_markdown"] = df_gpt4all["prompt"].apply(_markdownify_if_possible)
df_gpt4all["response_markdown"] = df_gpt4all["response"].apply(_markdownify_if_possible)
df_gpt4all = df_gpt4all.loc[df_gpt4all["prompt_markdown"].apply(bool) & df_gpt4all["response_markdown"].apply(bool)]
df_gpt4all.head()

df_gpt4all_texts = []
gpt4_all_indices = []
for _, row in df_gpt4all.iterrows():
    df_gpt4all_texts.append({"role": "prompter", "text": row["prompt_markdown"]})
    df_gpt4all_texts.append({"role": "assistant", "text": row["response_markdown"]})
    gpt4_all_indices.append([len(df_gpt4all_texts) - 2, len(df_gpt4all_texts) - 1])
df_gpt4all_texts = pd.DataFrame.from_records(df_gpt4all_texts)

gpt4all_indices_train, gpt4all_indices_validation = train_test_split(
    gpt4_all_indices,
    random_state=RANDOM_STATE,
    test_size=0.05
)

df_gpt4all_texts.loc[gpt4all_indices_train[0]]

### BookSum

dataset_booksum = load_dataset("kmfoda/booksum")
dataset_booksum

df_booksum_train = pd.concat([pd.DataFrame(dataset_booksum["train"]), pd.DataFrame(dataset_booksum["test"])]).reset_index(drop=True)
df_booksum_train.head()

df_booksum_validation = pd.DataFrame(dataset_booksum["validation"])
df_booksum_validation.head()

def booksum_process(df):
    result_df = []
    result_indices = []
    for _, row in df.iterrows():
        chapter = row["chapter"].strip()
        summary = json.loads(row["summary"])["summary"].strip()
        result_df.append({
            "role": "prompter",
            "text": f"{chapter}\n\nGive me the summary of previous text",
        })
        result_df.append({
            "role": "assistant",
            "text": summary,
        })
        result_indices.append([len(result_df) - 2, len(result_df) - 1])
    return pd.DataFrame.from_records(result_df), result_indices

df_booksum_texts_train, booksum_indices_train = booksum_process(df_booksum_train)

df_booksum_texts_validation, booksum_indices_validation = booksum_process(df_booksum_validation)

df_booksum_texts_train.loc[booksum_indices_train[0]]

len(booksum_indices_train)

### GovReport

dataset_govreport = load_dataset("ccdv/govreport-summarization")
dataset_govreport

df_govreport_train = pd.concat([pd.DataFrame(dataset_govreport["train"]), pd.DataFrame(dataset_govreport["test"])]).reset_index(drop=True)
df_govreport_validation = pd.DataFrame(dataset_govreport["validation"])

def govreport_process(df):
    result_df = []
    result_indices = []
    for _, row in df.iterrows():
        report = row["report"].strip()
        summary = row["summary"].strip()
        result_df.append({
            "role": "prompter",
            "text": f"{report}\n\nGive me the summary of previous text",
        })
        result_df.append({
            "role": "assistant",
            "text": summary,
        })
        result_indices.append([len(result_df) - 2, len(result_df) - 1])
    return pd.DataFrame.from_records(result_df), result_indices

df_govreport_texts_train, govreport_indices_train = govreport_process(df_govreport_train)
df_govreport_texts_validation, govreport_indices_validation = govreport_process(df_govreport_validation)

df_govreport_texts_train.loc[govreport_indices_train[0]]

### QASper

dataset_qasper = load_dataset("allenai/qasper")
dataset_qasper

def _qasper_article_text(data):
    section_texts = []
    for section_name, section_paragraphs in zip(data['section_name'], data['paragraphs']):
        if not section_name:
            section_name = ""
        section_name_levels = section_name.split(" ::: ")
        section_name_last = section_name_levels[-1]
        if section_name:
            section_name_prefix = "#" * len(section_name_levels)
            title = f"{section_name_prefix} {section_name_last}"
        else:
            title = ""
        section_content = "\n\n".join(section_paragraphs)
        section_texts.append(f"{title}\n\n{section_content}")
    return "\n\n".join(section_texts)

def _qasper_article_qa_pairs(full_text_data, qas_data):
    article_text = _qasper_article_text(full_text_data)

    question_texts = qas_data["question"]
    answers = qas_data["answers"]
    for question, question_answers in zip(question_texts, answers):
        for answer in question_answers["answer"]:
            answer_evidence = answer["highlighted_evidence"]
            answer_text = answer["free_form_answer"]
            if answer_evidence:
                answer_evidence_text = "\n".join([f"- {item}" for item in answer_evidence])
                answer_message_text = f"{answer_evidence_text}\n\n{answer_text}"
            else:
                answer_message_text = answer_text
            yield article_text, f"Keeping this in mind, answer the following question: {question}", answer_message_text.strip()

def quasper_qa_pairs(df):
    texts = []
    indices = []
    for _, row in df.iterrows():
        row_qa_texts = []
        for article_text, question_text, answer in _qasper_article_qa_pairs(row["full_text"], row["qas"]):
            row_qa_texts.append((question_text, answer))
        
        texts.append({"role": "prompter", "text": article_text})
        index_article = len(texts) - 1
        for question, answer in row_qa_texts:
            texts.append({"role": "prompter", "text": question})
            texts.append({"role": "assistant", "text": answer})
            index_question = len(texts) - 2
            index_answer = len(texts) - 1
            indices.append([index_article, index_question, index_answer])
    
    return pd.DataFrame.from_records(texts), indices

df_qasper_train = pd.concat([pd.DataFrame(dataset_qasper["train"]), pd.DataFrame(dataset_qasper["test"])]).reset_index(drop=True)
df_qasper_validation = pd.DataFrame(dataset_qasper["validation"])

df_qasper_texts_train, qasper_indices_train = quasper_qa_pairs(df_qasper_train)
df_qasper_texts_validation, qasper_indices_validation = quasper_qa_pairs(df_qasper_validation)

df_qasper_texts_train.loc[qasper_indices_train[0]]

print(df_qasper_texts_train.loc[qasper_indices_train[0]]["text"].values[-1])

print(df_qasper_texts_validation.loc[qasper_indices_validation[0]]["text"].values[-1])

print(df_qasper_texts_validation.loc[qasper_indices_validation[10]]["text"].values[-1])

len(qasper_indices_train)

## Combine sets

def shift_indices(indices, offset):
    return [
        [i + offset for i in row]
        for row in indices
    ]

texts = []
indices_train = []
indices_validation = []
sources_train = []
sources_validation = []

indices_train.append(shift_indices(
    oasst1_indices_train,
    sum(map(len, texts))
))
texts.append(df_oasst1_texts_train)
sources_train.append("openassistant")

indices_validation.append(shift_indices(
    oasst1_indices_validation,
    sum(map(len, texts))
))
texts.append(df_oasst1_texts_validation)
sources_validation.append("openassistant")

indices_train.append(shift_indices(
    alpaca_indices_train,
    sum(map(len, texts))
))
sources_train.append("alpaca")

indices_validation.append(shift_indices(
    alpaca_indices_validation,
    sum(map(len, texts))
))
sources_validation.append("alpaca")

texts.append(df_alpaca_texts)

indices_train.append(shift_indices(
    gpt4all_indices_train,
    sum(map(len, texts))
))
sources_train.append("gpt4all")

indices_validation.append(shift_indices(
    gpt4all_indices_validation,
    sum(map(len, texts))
))
sources_validation.append("gpt4all")

texts.append(df_gpt4all_texts)

indices_train.append(shift_indices(
    booksum_indices_train,
    sum(map(len, texts))
))
texts.append(df_booksum_texts_train)
sources_train.append("booksum")

indices_validation.append(shift_indices(
    booksum_indices_validation,
    sum(map(len, texts))
))
texts.append(df_booksum_texts_validation)
sources_validation.append("booksum")

indices_train.append(shift_indices(
    govreport_indices_train,
    sum(map(len, texts))
))
texts.append(df_govreport_texts_train)
sources_train.append("govreport")

indices_validation.append(shift_indices(
    govreport_indices_validation,
    sum(map(len, texts))
))
texts.append(df_govreport_texts_validation)
sources_validation.append("govreport")

indices_train.append(shift_indices(
    qasper_indices_train,
    sum(map(len, texts))
))
texts.append(df_qasper_texts_train)
sources_train.append("qasper")

indices_validation.append(shift_indices(
    qasper_indices_validation,
    sum(map(len, texts))
))
texts.append(df_qasper_texts_validation)
sources_validation.append("qasper")

def get_indices_df(indices, sources):
    rows = []
    for ds_indices, ds_source in zip(indices, sources):
        for row in ds_indices:
            rows.append({"indices": row, "source": ds_source})
    return pd.DataFrame.from_records(rows)

df_indices_train = get_indices_df(indices_train, sources_train)
df_indices_train.head()

df_indices_validation = get_indices_df(indices_validation, sources_validation)
df_indices_validation.head()

df_texts = pd.concat(texts).reset_index(drop=True)
df_texts.head()

df_indices_train.groupby("source").head(1)

pd.set_option("display.max_colwidth", 256)

df_texts.loc[[0, 1, 2]]

df_texts.loc[[139334, 139335]]

df_texts.loc[[825928, 825929]]

df_texts.loc[[958912, 958913]]

df_texts.loc[[983942, 983943]]

df_texts.loc[[1022868, 1022869, 1022870]]

df_indices_validation.groupby("source").head(1)

df_texts.loc[[82483, 82484]]

df_texts.loc[[163170, 163171]]

df_texts.loc[[705722, 705723]]

df_texts.loc[[980974, 980975]]

df_texts.loc[[1020922, 1020923]]

df_texts.loc[[1036630, 1036631, 1036632]]

os.makedirs("long-vicuna-set", exist_ok=True)
df_texts.to_csv("long-vicuna-set/texts.gz", index=True, compression="gzip")
df_indices_train.to_pickle("long-vicuna-set/indices-train.pkl")
df_indices_validation.to_pickle("long-vicuna-set/indices-validation.pkl")
