==> nbs/06_train_longvicuna__preprocess_data.ipynb <==
# Dataset preprocessing

import pandas as pd
from itertools import chain
from llama_4bit_wrapper import import_llama, lora_model_zeros_and_scales_to_half
from tqdm import tqdm
import numpy as np
import os

tqdm.pandas()

RANDOM_STATE = 42

### Loading pre-trained model

_, _, load_llama_model_4bit_low_ram, _, _, _, apply_gradient_checkpointing, _, _ = import_llama(
    use_flash_attention=False,
    use_xformers=False,
    autograd_4bit_cuda=False,
    autograd_4bit_triton=True,
)

_, tokenizer = load_llama_model_4bit_low_ram(
    config_path="../vicuna-13b-GPTQ-4bit-128g/",
    model_path="../vicuna-13b-GPTQ-4bit-128g/vicuna-13b-4bit-128g.safetensors",
    groupsize=128,
    is_v1_model=False,
)
tokenizer.pad_token_id = 0

### Loading data

df_texts = pd.read_csv("long-vicuna-set/texts.gz", compression="gzip", index_col=0)
df_texts.head()

df_indices_train = pd.read_pickle("long-vicuna-set/indices-train.pkl")
df_indices_train.head()

df_indices_validation = pd.read_pickle("long-vicuna-set/indices-validation.pkl")
df_indices_validation.head()

df_indices_train["source"].value_counts()

df_indices_train["source"].value_counts() / df_indices_train.shape[0]

df_indices_validation["source"].value_counts()

df_indices_validation["source"].value_counts() / df_indices_validation.shape[0]

### Reducing gpt4all usage

Since GPT4All is a syntetic dataset - let's reduce it's usage. It's good for "let's follow this instruction format" training but we can't be sure about the quality.

def _get_dataset_fraction(df, source, fraction, random_state):
    df_filtered = df.loc[df["source"] == source]
    df_filtered = df_filtered.sample(int(df_filtered.shape[0] * fraction), random_state=random_state)
    return df_filtered

def _limit_dataset_fraction(df, source, fraction, random_state):
    df_filtered = df.loc[df["source"] != source]
    df = pd.concat([df_filtered, _get_dataset_fraction(df, source, fraction, random_state)])
    return df.loc[sorted(df.index)]

df_indices_train = _limit_dataset_fraction(df_indices_train, "gpt4all", 0.1, RANDOM_STATE)
df_indices_validation = _limit_dataset_fraction(df_indices_validation, "gpt4all", 0.1, RANDOM_STATE)

df_indices_train["source"].value_counts()

df_indices_validation["source"].value_counts()

### Removing unnecessary texts

text_indices_to_process = sorted(set(chain(*df_indices_train["indices"], *df_indices_validation["indices"])))
len(df_texts), len(text_indices_to_process)

df_texts = df_texts.loc[text_indices_to_process]
df_texts.head()

### Tokenization

tokenizer.convert_ids_to_tokens(
    tokenizer("TEST")["input_ids"]
)

tokenizer.vocab_size

2 ** 16

df_texts["processed_text"] = "<msg_" + df_texts["role"] + "> " + df_texts["text"].fillna("").str.strip()
df_texts = df_texts[["processed_text"]]

df_texts["input_ids"] = df_texts["processed_text"].fillna("").progress_apply(
    lambda text: np.array(tokenizer(text)["input_ids"][1:], dtype=np.uint16),
)

df_texts.head()

### Adding lengths

df_texts["length"] = df_texts["input_ids"].apply(len)
df_texts.head()

(df_texts["processed_text"].str.len() / df_texts["length"]).mean()

df_indices_train["session_length"] = df_indices_train["indices"].apply(lambda indices: df_texts["length"][indices].sum())
df_indices_validation["session_length"] = df_indices_validation["indices"].apply(lambda indices: df_texts["length"][indices].sum())

df_indices_train.head()

df_indices_validation.head()

df_indices_train.groupby("source")["session_length"].mean()

df_indices_train.groupby("source")["session_length"].quantile(0.9)

### Saving the processed set

os.makedirs("long-vicuna-set-lessgpt4all-vicuna13b-processed", exist_ok=True)
df_texts.to_pickle("long-vicuna-set-lessgpt4all-vicuna13b-processed/texts.pkl")
df_indices_train.to_pickle("long-vicuna-set-lessgpt4all-vicuna13b-processed/indices-train.pkl")
df_indices_validation.to_pickle("long-vicuna-set-lessgpt4all-vicuna13b-processed/indices-validation.pkl")
