==> nbs/07_train_longvicuna__pretrain_loop.ipynb <==
import os
import pandas as pd
from llama_4bit_wrapper import import_llama, lora_model_zeros_and_scales_to_half
from peft import LoraConfig, get_peft_model
from llama_memorizing_transformers.memory_collection import CosineKnnMemoryCollection
from llama_memorizing_transformers.context_choice import ContextChoiceLinear
from llama_memorizing_transformers.model_wrapper import replace_llama_layer_with_memory
from llama_memorizing_transformers.document_trainer import MemorizingLlamaDocumentTrainer
from torch.optim import AdamW
from transformers import get_constant_schedule_with_warmup
import torch
import numpy as np
from torch.utils.tensorboard import SummaryWriter
from tqdm import tqdm

RANDOM_STATE = 42

# Dataset
DATASET_PATH = "long-vicuna-set-lessgpt4all-vicuna13b-processed"

# Training procedure
CONTEXT_LENGTH = 512
CONTEXT_STEP = 256
PRETRAIN_LENGTH = 2048
PRETRAIN_DOCUMENTS = 2048
WARMUP_STEPS = PRETRAIN_DOCUMENTS * ((PRETRAIN_LENGTH - CONTEXT_LENGTH) / CONTEXT_STEP + 1)

# Model
COSINE_KNN_MAX_TEMPORARY_BUFFER_SIZE = 1024
REPLACE_LAYER = 21
BASE_MODEL = "../vicuna-13b-GPTQ-4bit-128g"
BASE_MODEL_WEIGHTS = "../vicuna-13b-GPTQ-4bit-128g/vicuna-13b-4bit-128g.safetensors"

USE_FP16 = True
LR_PRETRAIN = 3e-4

_, _, load_llama_model_4bit_low_ram, _, _, _, apply_gradient_checkpointing, _, _ = import_llama(
    use_flash_attention=False,
    use_xformers=False,
    autograd_4bit_cuda=False,
    autograd_4bit_triton=True,
)

## Data reading

df_texts = pd.read_pickle(os.path.join(DATASET_PATH, "texts.pkl"))
df_texts.head()

df_indices_train = pd.read_pickle(os.path.join(DATASET_PATH, "indices-train.pkl"))
df_indices_train.head()

df_indices_validation = pd.read_pickle(os.path.join(DATASET_PATH, "indices-validation.pkl"))
df_indices_validation.head()

## Model preparation

model, tokenizer = load_llama_model_4bit_low_ram(
    config_path=BASE_MODEL,
    model_path=BASE_MODEL_WEIGHTS,
    groupsize=128,
    is_v1_model=False,
)
model.cuda()
tokenizer.pad_token_id = 0

context_choice = ContextChoiceLinear(model.config.num_attention_heads,
                                     model.config.hidden_size)
context_choice.cuda()

memory = CosineKnnMemoryCollection(COSINE_KNN_MAX_TEMPORARY_BUFFER_SIZE,
                                   remember_until_position=0)

model.model = replace_llama_layer_with_memory(
    model.model,
    REPLACE_LAYER,
    context_choice,
    memory,
)

lora_config = LoraConfig(
    r=8,
    lora_alpha=16,
    target_modules=["q_proj", "v_proj"],
    lora_dropout=0.0,
    bias="none",
    task_type="CAUSAL_LM",
)
lora_model = get_peft_model(model, lora_config)
lora_model = lora_model_zeros_and_scales_to_half(lora_model)

apply_gradient_checkpointing(lora_model)

lora_model.config.use_cache = False

## Tokenized texts checking

df_texts.sample(5, random_state=RANDOM_STATE)

df_texts.sample(5, random_state=RANDOM_STATE)["input_ids"].apply(
    lambda item: tokenizer.decode([tokenizer.bos_token_id] + list(item) + [tokenizer.eos_token_id])
)

df_indices_train.loc[df_indices_train["session_length"] <= PRETRAIN_LENGTH].groupby("source")["session_length"].quantile(0.5)

### Pretraining

optimizer = AdamW(
    lora_model.parameters(),
    lr=LR_PRETRAIN,
)
scheduler = get_constant_schedule_with_warmup(optimizer, WARMUP_STEPS)

df_indices_pretrain = df_indices_train\
    .loc[(df_indices_train["session_length"] <= PRETRAIN_LENGTH) & (df_indices_train["session_length"] >CONTEXT_LENGTH)]\
    .sample(PRETRAIN_DOCUMENTS, random_state=RANDOM_STATE)
df_indices_pretrain.head()

_batch_counter = 0

def batch_counter_update() -> int:
    global _batch_counter
    _batch_counter += 1
    return _batch_counter

def build_train_callback(log_writer):
    def train_callback(loss, loss_lm, loss_context, **kwargs):
        index = batch_counter_update()
        log_writer.add_scalar("Loss/pretrain", loss, index)
        log_writer.add_scalar("LossLM/pretrain", loss_lm, index)
        log_writer.add_scalar("LossContext/pretrain", loss_context, index)

    return train_callback

lora_model.base_model.model.model.layers[20]

log_writer = SummaryWriter("long-vicuna--pretrain--tensorboard")
document_trainer = MemorizingLlamaDocumentTrainer(
    model=lora_model,
    context_choice=context_choice,
    tokenizer=tokenizer,
    memory=memory,
    tokens_per_chunk=CONTEXT_LENGTH,
    tokens_step=CONTEXT_STEP,
    optimizer=optimizer,
    scheduler=scheduler,
    accumulate_gradients=1,
    float16=USE_FP16,
    train_callback=build_train_callback(log_writer),
    eval_callback=None,
)
for i, indices in enumerate(tqdm(df_indices_pretrain["indices"])):
    main_document_index = indices[0]
    rest_session_index = indices[1:]
    document_tokens = torch.LongTensor(df_texts.loc[main_document_index, "input_ids"].astype(np.int32))
    rest_session_tokens = torch.cat([
        torch.LongTensor(array.astype(np.int32))
        for array in df_texts.loc[rest_session_index, "input_ids"]
    ]).view((1, -1))
    document_trainer.train_document(
        document_tokens=document_tokens,
        prompt_tokens=rest_session_tokens,
        sample_weight=1.0,
        callback_kwargs={
            "document_index": i,
        }
    )
    if ((i % 500) == 0) and (i > 0):
        torch.save(lora_model.state_dict(), f"long-vicuna--pretrain--state-dict--checkpoint-{i}.pth")
